requirements.txt
alpha_vantage==2.3.1
matplotlib==3.7.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6